Move E2eTestHelpers -> E2e::TestHelpers
