+++ app/spec/e2e/cbv_flow_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe "e2e CBV flow test", type: :feature, js: true do
-  include E2eTestHelpers
+  include E2e::TestHelpers
 
   let(:cbv_flow_invitation) { create(:cbv_flow_invitation) }
